+++ database/migrations/2022_02_14_225416_create_reservations_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('reservations', function (Blueprint $table) {
             $table->id();
-            $table->integer('shop_id');
+            $table->unsignedBigInteger('shop_id');
             $table->foreign('shop_id')->references('id')->on('shops');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreignId('user_id')->references('id')->on('users');
             $table->string('num_of_users')->nullable(false);
             $table->date('date')->nullable(false);
